some clenaup
